requirements.txt
pytesseract
pdf2image
python-dotenv
pygtrie
pandas
numpy

